More robust marker characters
This improves the marker characters so things work, even if the marker
character is used in the text itself.

This is necessary, because we want to use space as the marker character
to get proper capitalization on iOS.

To make the marker characters invisible, SpannedTextEditingController is
made aware of them. It then replaces the characters with ZWSP when
rendering in `buildTextSpan`. This keeps the offsets for the span correct,
but does not render a visible character.

This mode is switched with a flag: `hasMarkerCharacters`.